Add provider, and a main routes files
