+++ App.js
@@ -1,44 +1,15 @@
+import 'react-native-gesture-handler';
 import React from 'react';
-import { createStackNavigator } from "@react-navigation/stack";
-import { NavigationContainer, DefaultTheme } from '@react-navigation/native';
-
-import { Home } from "./screens/";
-import { useFonts } from 'expo-font';
-
-const theme = {
-    ...DefaultTheme,
-    colors: {
-        ...DefaultTheme.colors,
-        border: "transparent",
-    },
-};
-
-const Stack = createStackNavigator();
+import { store } from './redux/Store';
+import { Provider } from 'react-redux';
+import Routes from './Routes';
 
 const App = () => {
-
-    const [loaded] = useFonts({
-        "Roboto-Black" : require('./assets/fonts/Roboto-Black.ttf'),
-        "Roboto-Bold" : require('./assets/fonts/Roboto-Bold.ttf'),
-        "Roboto-Regular" : require('./assets/fonts/Roboto-Regular.ttf'),
-    })
-
-    if(!loaded){
-        return null;
-    }
-    
-    return (
-        <NavigationContainer theme={theme}>
-            <Stack.Navigator
-                screenOptions={{
-                    headerShown: false
-                }}
-                initialRouteName={'Home'}
-            >
-                <Stack.Screen name="Home" component={Home} />
-            </Stack.Navigator>
-        </NavigationContainer>
-    );
+  return (
+    <Provider store={store}>
+      <Routes />
+    </Provider>
+  );
 };
 
 export default App;
